CMS7-7941 Add expand/collapse functionality
If a parent is expanded (collapsed is false) then all child nodes will be rendered.
In the theme demo I have added a button to togggle the expand/collapse flag of a node.
The button is visible if the node has children.
The button shows a + if it's node is collapsed and a - if it is expanded.