CMS7-7954 Synchronize model and ui-tree collapsed state
To be able to restore the collapsed state of a tree the tree controller has been augmented
with a toggleItem function that will keep the ui-tree and model collapsed state in sync.
By watching the model internally, this information is used to restore the collapsed state
after a reload of the model.