CMS7-7915 Hide IE11 pseudo elements on input fields
As of IE10, an X button is rendered by default inside of an input field.
Clicking this button clears the input field. For password fields an icon
is displayed that enables to reveal the password. Although useful, the
result is that the application behaves differently than in other browsers.
So to enforce consistent behavior rendering of pseudo elements is suppressed.